COMMERCE-960 commerce-openapi - refactor, move LanguageUtils to openapi core module so other modules can use it
