+++ commerce-openapi-admin/src/main/java/com/liferay/commerce/openapi/admin/internal/util/v2_0/DTOUtils.java
@@ -19,7 +19,6 @@ import com.liferay.commerce.currency.model.CommerceCurrency;
 import com.liferay.commerce.model.CommerceCountry;
 import com.liferay.commerce.model.CommerceWarehouse;
 import com.liferay.commerce.model.CommerceWarehouseItem;
-import com.liferay.commerce.openapi.admin.internal.util.LanguageUtils;
 import com.liferay.commerce.openapi.admin.model.v2_0.AccountDTO;
 import com.liferay.commerce.openapi.admin.model.v2_0.CountryDTO;
 import com.liferay.commerce.openapi.admin.model.v2_0.CurrencyDTO;
@@ -32,6 +31,7 @@ import com.liferay.commerce.openapi.admin.model.v2_0.ProductOptionValueDTO;
 import com.liferay.commerce.openapi.admin.model.v2_0.SkuDTO;
 import com.liferay.commerce.openapi.admin.model.v2_0.UserDTO;
 import com.liferay.commerce.openapi.admin.model.v2_0.WebSiteDTO;
+import com.liferay.commerce.openapi.core.util.LanguageUtils;
 import com.liferay.commerce.price.list.model.CommercePriceEntry;
 import com.liferay.commerce.price.list.model.CommercePriceList;
 import com.liferay.commerce.product.model.CPDefinition;
